+++ source/extensions/isaacsim.core.simulation_manager/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [1.14.3] - 2026-04-17
+### Changed
+- Update `PhysxScene` to use renamed PhysX attribute `GpuMaxDeformableVolumeContacts`
+
 ## [1.14.2] - 2026-04-07
 ### Added
 - Add optional multitick support to TimeSampleStorage and PluginInterface. When multitick is enabled, physics time drives rendering time via onPhysicsStep callback.